Net 8.0 BFF + PAR Example
While this example is not BFF, I was able to use this implementation in BFF with minimal changes. 

Perhaps this could use a whole new example or is it not even worth it since from [.net9-preview7](https://github.com/dotnet/core/blob/main/release-notes/9.0/preview/preview7/aspnetcore.md#openidconnecthandler-support-for-pushed-authorization-requests-par) it works out of the box anyway.